feat: add OdooGhost entry point file
Add a new file `__main__.py` in the `odooghost` directory to serve as the entry point for the OdooGhost application.

feat: add main module for CLI

Add a new file `main.py` in the `odooghost/cli` directory to define the main module for the OdooGhost command-line interface (CLI).

chore: update project version

Update the version of the project in the `pyproject.toml` file from `0.1.0-alpha1` to `0.1.0`.

feat: add typer dependency

Add the `typer` dependency to the project in the `pyproject.toml` file with a version constraint of `^0.9.0`.

feat: add odooghost script

Add a script entry for `odooghost` in the `pyproject.toml` file to run the `odooghost.cli.main:cli` command as the entry point for the `odooghost` script.