feat(services): add new services module with base, odoo, and postgres modules
The new services module is added to the odooghost package. It includes three modules: base, odoo, and postgres.

The base module contains the abstract base class `BaseService` which defines the common functionality for all services. It includes a method `ensure_image()` to ensure that the required Docker image is available. The `image_tag` property is defined as an abstract property.

The odoo module contains the `OdooService` class which extends `BaseService`. It takes a `config` parameter of type `stack.OdooStackConfig` and sets it as an instance variable. The `image_tag` property is implemented to return the image tag based on the version specified in the config.

The postgres module contains the `PostgresService` class which also extends `BaseService`. It takes a `config` parameter of type `stack.PostgresStackConfig` and sets it as an instance variable. It includes an `is_remote` property to check if the type of the config is "remote". The `image_tag` property is implemented to return the image tag based on the version specified in the config.