chore(orchestrator.yaml): update file references to use correct file extensions
The file references in the orchestrator.yaml file have been updated to use the correct file extensions. The references to test.yml and lint.yml have been changed to test.yaml and lint.yaml respectively. Similarly, the reference to release.yml has been changed to release.yaml. This change ensures that the correct workflow files are used in the respective jobs.