chore(release.yaml): update Python Semantic Release step to continue on error and add Publish step to handle failure case
The Python Semantic Release step in the release.yaml workflow has been updated to include the `continue-on-error: true` option. This change allows the workflow to continue even if the semantic-release command fails.

Additionally, a new Publish step has been added to the workflow. This step is only executed if the previous Python Semantic Release step fails. It runs the `poetry publish -n` command to publish the package to the package index.