fix(container.py): fix return type annotation for the get_archive method in the Container class
The return type annotation for the get_archive method in the Container class was incorrect. It was originally annotated as returning a tuple of t.IO and t.Dict[str, t.Any], but it should be annotated as returning a tuple of t.Generator[any, any, None] and t.Dict[str, t.Any]. This fix improves the accuracy of the return type annotation for the method.