feat(config.py): add support for mapping local port to container service port in PostgresStackConfig and OdooStackConfig classes
The changes were made to the `config.py` file. In the `PostgresStackConfig` and `OdooStackConfig` classes, a new attribute `service_port` of type `Optional[int]` was added. This attribute allows mapping a local port to the container service port.

Additionally, in the `DependenciesConfig` class, the `get_apt_dependencies` method was updated to handle both string and list inputs. If the input is a string, it is split by spaces to create a list of apt dependencies.

In the `OdooStackConfig` class, a validator was added to validate the `version` attribute as a float.