fix(addons.py): add a pull method to pull remote Odoo addons
The `AddonsHandler` class in `addons.py` now has a new method called `pull` which is used to pull remote Odoo addons. This method takes an optional argument `depth` which specifies the depth of the git pull. The method iterates over the addons and if an addon is of type "remote", it performs a git pull on the addon's path using the specified branch and depth. This allows for easily updating remote addons without having to manually clone the repository again.