feat(StackView.vue): add fetchPolicy and pollInterval options to useQuery function call to improve data fetching behavior
The changes were made to the `StackView.vue` file. The `useQuery` function call was modified to include the `fetchPolicy` and `pollInterval` options.

The `fetchPolicy` option was set to `'cache-and-network'` to ensure that the query fetches data from both the cache and the network. This allows for more up-to-date data to be displayed in the view.

The `pollInterval` option was set to `5000` to specify that the query should be refetched every 5 seconds. This enables the view to automatically update with the latest data without requiring manual refreshes.